backend/requirements.txt
Django~=3.2
psycopg2~=2.9
requests~=2.25
djangorestframework~=3.12